+++ packages/react-widgets/__tests__/models/FormulaModel.test.js
@@ -1,5 +1,5 @@
 import { minify } from 'pgsql-minify';
-import { getFormula, buildSqlQueryToGetFormula } from '../../src/models/FormulaModel';
+import { getFormula } from '../../src/models/FormulaModel';
 import { AggregationTypes } from '@carto/react-core';
 import { SourceTypes } from '@carto/react-api';
 import { Methods, executeTask } from '@carto/react-workers';
@@ -19,7 +19,7 @@ describe('getFormula', () => {
     );
   });
 
-  test('should throw if using CartoBQTilerLayer without viewportFilter', async () => {
+  test.skip('should throw if using CartoBQTilerLayer without viewportFilter', async () => {
     await expect(
       getFormula({ type: SourceTypes.BIGQUERY, viewportFilter: false })
     ).rejects.toThrow(
@@ -27,7 +27,32 @@ describe('getFormula', () => {
     );
   });
 
-  describe('SQL Layer', () => {
+  describe('should correctly handle viewport features', () => {
+    const formulaParams = {
+      operation: AggregationTypes.COUNT,
+      column: 'revenue',
+      filters: {},
+      dataSource: 'whatever-data-source'
+    };
+
+    test('correctly returns data', async () => {
+      executeTask.mockImplementation(() => Promise.resolve([{ value: 3 }]));
+      const formula = await getFormula(formulaParams);
+      expect(formula).toEqual([{ value: 3 }]);
+    });
+
+    test('correctly called', async () => {
+      const { operation, column, filters, dataSource } = formulaParams;
+      await getFormula(formulaParams);
+      expect(executeTask).toHaveBeenCalledWith(
+        dataSource,
+        Methods.VIEWPORT_FEATURES_FORMULA,
+        { column, filters, operation }
+      );
+    });
+  });
+
+  describe.skip('SQL Layer', () => {
     describe('should execute an SqlApi request when using "viewportFilter": false', () => {
       const response = { rows: { revenue: 1495728 } };
       const sql = 'SELECT revenue FROM retail_stores LIMIT 1';
@@ -56,50 +81,5 @@ describe('getFormula', () => {
         expect(func).toEqual(response.rows);
       });
     });
-
-    describe('should handle viewport features when using "viewportFilter": true', () => {
-      const viewportFeatures = [...Array(3)].map((_, idx) => ({ revenue: idx }));
-
-      const formulaParams = {
-        operation: AggregationTypes.COUNT,
-        column: 'revenue',
-        type: SourceTypes.SQL,
-        viewportFilter: true,
-        viewportFeatures,
-        dataSource: 'whatever-data-source',
-        filters: {}
-      };
-
-      test('correctly returns data', async () => {
-        executeTask.mockImplementation(() => Promise.resolve([{ value: 3 }]));
-        const formula = await getFormula(formulaParams);
-        expect(formula).toEqual([{ value: 3 }]);
-      });
-
-      test('correctly called', async () => {
-        const { column, dataSource, filters, operation } = formulaParams;
-        await getFormula(formulaParams);
-        expect(executeTask).toHaveBeenCalledWith(
-          dataSource,
-          Methods.VIEWPORT_FEATURES_FORMULA,
-          { column, filters, operation }
-        );
-      });
-    });
-  });
-});
-
-describe('buildSqlQueryToGetFormula', () => {
-  test('should work as expected', () => {
-    const params = {
-      data: 'SELECT * FROM stores',
-      column: 'revenue',
-      operation: AggregationTypes.COUNT
-    };
-    const buildQuery = (type) =>
-      `SELECT ${type}(revenue) as value FROM (SELECT * FROM stores) as q`;
-    expect(buildSqlQueryToGetFormula(params)).toEqual(
-      minify(buildQuery(params.operation))
-    );
   });
 });