+++ packages/react-widgets/src/widgets/CategoryWidget.js
@@ -22,61 +22,63 @@ import useWidgetLoadingState from './useWidgetLoadingState';
  * @param  {string} props.operation - Operation to apply to the operationColumn. Must be one of those defined in `AggregationTypes` object.
  * @param  {formatterCallback} [props.formatter] - Function to format each value returned.
  * @param  {Object} [props.labels] - Overwrite category labels
- * @param  {boolean} [props.viewportFilter=true] - Defines whether filter by the viewport or globally.
  * @param  {errorCallback} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
  */
 function CategoryWidget(props) {
-  const { column } = props;
+  const {
+    id,
+    title,
+    dataSource,
+    column,
+    operationColumn,
+    operation,
+    formatter,
+    labels,
+    onError,
+    wrapperProps
+  } = props;
   const [categoryData, setCategoryData] = useState(null);
   const [selectedCategories, setSelectedCategories] = useState([]);
   const dispatch = useDispatch();
-  const source = useSelector((state) => selectSourceById(state, props.dataSource) || {});
+  const source = useSelector((state) => selectSourceById(state, dataSource) || {});
   const viewportFeaturesReady = useSelector((state) => state.carto.viewportFeaturesReady);
   const widgetsLoadingState = useSelector((state) => state.carto.widgetsLoadingState);
-  const [hasLoadingState, setIsLoading] = useWidgetLoadingState(
-    props.id,
-    props.viewportFilter
-  );
-  const { data, credentials, type } = source;
+  const [hasLoadingState, setIsLoading] = useWidgetLoadingState(id);
+  const { data, filters } = source;
 
   useEffect(() => {
-    const abortController = new AbortController();
-    if (data && credentials && hasLoadingState) {
-      const filters = getApplicableFilters(source.filters, props.id);
-      !props.viewportFilter && setIsLoading(true);
+    if (data && hasLoadingState) {
+      const _filters = getApplicableFilters(filters, id);
       getCategories({
-        ...props,
         data,
-        filters,
-        credentials,
-        viewportFeatures: viewportFeaturesReady[props.dataSource] || false,
-        dataSource: props.dataSource,
-        type,
-        opts: { abortController }
+        column,
+        operationColumn,
+        operation,
+        filters: _filters,
+        dataSource
       })
         .then((data) => setCategoryData(data))
         .catch((error) => {
           if (error.name === 'AbortError') return;
-          if (props.onError) props.onError(error);
+          if (onError) onError(error);
         })
         .finally(() => setIsLoading(false));
     } else {
       setCategoryData(null);
     }
-
-    return function cleanup() {
-      abortController.abort();
-    };
   }, [
-    credentials,
+    id,
     data,
-    setIsLoading,
-    source.filters,
-    type,
+    column,
+    operationColumn,
+    operation,
+    filters,
+    dataSource,
     viewportFeaturesReady,
-    props,
-    hasLoadingState
+    setIsLoading,
+    hasLoadingState,
+    onError
   ]);
 
   const handleSelectedCategoriesChange = useCallback(
@@ -86,36 +88,32 @@ function CategoryWidget(props) {
       if (categories && categories.length) {
         dispatch(
           addFilter({
-            id: props.dataSource,
+            id: dataSource,
             column,
             type: FilterTypes.IN,
             values: categories,
-            owner: props.id
+            owner: id
           })
         );
       } else {
         dispatch(
           removeFilter({
-            id: props.dataSource,
+            id: dataSource,
             column
           })
         );
       }
     },
-    [column, props.dataSource, props.id, setSelectedCategories, dispatch]
+    [column, dataSource, id, setSelectedCategories, dispatch]
   );
 
   return (
-    <WrapperWidgetUI
-      title={props.title}
-      isLoading={widgetsLoadingState[props.id]}
-      {...props.wrapperProps}
-    >
+    <WrapperWidgetUI title={title} isLoading={widgetsLoadingState[id]} {...wrapperProps}>
       <CategoryWidgetUI
         data={categoryData}
-        formatter={props.formatter}
-        labels={props.labels}
-        isLoading={widgetsLoadingState[props.id]}
+        formatter={formatter}
+        labels={labels}
+        isLoading={widgetsLoadingState[id]}
         selectedCategories={selectedCategories}
         onSelectedCategoriesChange={handleSelectedCategoriesChange}
       />
@@ -132,14 +130,12 @@ CategoryWidget.propTypes = {
   operation: PropTypes.oneOf(Object.values(AggregationTypes)).isRequired,
   formatter: PropTypes.func,
   labels: PropTypes.object,
-  viewportFilter: PropTypes.bool,
   onError: PropTypes.func,
   wrapperProps: PropTypes.object
 };
 
 CategoryWidget.defaultProps = {
   labels: {},
-  viewportFilter: true,
   wrapperProps: {}
 };